moved all tests into a single tests directory
